Update README.md
Feat: Remove the code and attach file instead

Previously, the body contained a cut and paste of Python code of the project.  I want to use GitHub to submit the project but the reviewer was not able to review it. I am removing the code body and will submit code in attached file.